Adapt the public RecordBatchReader API to upstream's result_set refactor: add DatabaseDriverV1::result_set_get_reader, the non-FFI twin of result_set_get_stream, returning Box<dyn RecordBatchReader + Send + 'static> directly. Resolve conflicts in apis/database_driver_v1/{mod,query}.rs (upstream moved the query path into result_set) and regenerate the perf-app Cargo.lock.

+++ sf_core/src/rest/snowflake/query_request.rs
@@ -55,8 +55,8 @@ pub struct QueryContext {
 pub struct QueryContextEntry {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub context: Option<ContextData>,
-    pub id: i32,
-    pub priority: i32,
+    pub id: i64,
+    pub priority: i64,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub timestamp: Option<i64>,
 }
